# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonPrimitive"
            },
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonPrimitive": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "StudyFlashcardItemResponse": {
        "properties": {
          "back": {
            "title": "Back",
            "type": "string"
          },
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "front": {
            "title": "Front",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "id",
          "course_id",
          "topic",
          "front",
          "back",
          "source",
          "created_at"
        ],
        "title": "StudyFlashcardItemResponse",
        "type": "object"
      },
      "StudyFlashcardRequest": {
        "properties": {
          "back": {
            "minLength": 1,
            "title": "Back",
            "type": "string"
          },
          "course_id": {
            "exclusiveMinimum": 0.0,
            "title": "Course Id",
            "type": "integer"
          },
          "front": {
            "minLength": 1,
            "title": "Front",
            "type": "string"
          },
          "source": {
            "default": "study",
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "topic": {
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "topic",
          "front",
          "back"
        ],
        "title": "StudyFlashcardRequest",
        "type": "object"
      },
      "StudyFlashcardResponse": {
        "properties": {
          "flashcard": {
            "$ref": "#/components/schemas/StudyFlashcardItemResponse"
          }
        },
        "required": [
          "flashcard"
        ],
        "title": "StudyFlashcardResponse",
        "type": "object"
      },
      "StudySessionCompleteRequest": {
        "properties": {
          "post_test_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Post Test Score",
            "type": "number"
          },
          "pre_test_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Pre Test Score",
            "type": "number"
          }
        },
        "required": [
          "pre_test_score",
          "post_test_score"
        ],
        "title": "StudySessionCompleteRequest",
        "type": "object"
      },
      "StudySessionCompletionResponse": {
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "session_id": {
            "title": "Session Id",
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "completed"
        ],
        "title": "StudySessionCompletionResponse",
        "type": "object"
      },
      "StudySessionItemResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "improvement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improvement"
          },
          "post_test_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Test Score"
          },
          "pre_test_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pre Test Score"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "id",
          "course_id",
          "topic",
          "pre_test_score",
          "post_test_score",
          "started_at",
          "completed_at",
          "improvement"
        ],
        "title": "StudySessionItemResponse",
        "type": "object"
      },
      "StudySessionListResponse": {
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/StudySessionItemResponse"
            },
            "title": "Sessions",
            "type": "array"
          }
        },
        "required": [
          "course_id",
          "sessions"
        ],
        "title": "StudySessionListResponse",
        "type": "object"
      },
      "StudySessionResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/StudySessionItemResponse"
          }
        },
        "required": [
          "session"
        ],
        "title": "StudySessionResponse",
        "type": "object"
      },
      "StudySessionStartRequest": {
        "properties": {
          "course_id": {
            "exclusiveMinimum": 0.0,
            "title": "Course Id",
            "type": "integer"
          },
          "topic": {
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "topic"
        ],
        "title": "StudySessionStartRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Sophia API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/study/flashcards": {
      "post": {
        "operationId": "saveStudyFlashcard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyFlashcardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyFlashcardResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Create Study Flashcard",
        "tags": [
          "study"
        ]
      }
    },
    "/api/study/sessions": {
      "get": {
        "operationId": "listStudySessions",
        "parameters": [
          {
            "in": "query",
            "name": "course_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Course Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySessionListResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Study Sessions",
        "tags": [
          "study"
        ]
      },
      "post": {
        "operationId": "startStudySession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Create Study Session",
        "tags": [
          "study"
        ]
      }
    },
    "/api/study/sessions/{session_id}/complete": {
      "post": {
        "operationId": "completeStudySession",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySessionCompletionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Mark Study Session Complete",
        "tags": [
          "study"
        ]
      }
    },
...
  }
}